pyproject.toml
[project]
name = "0810hackathon-back"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "python-dotenv",
    "uvicorn[standard]",
    "sqlalchemy",
    "pydantic-settings>=2.10.1",
    "psycopg2-binary>=2.9.10",
    "geopy>=2.4.1",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D1",    # undocumented
    "D203",  # one blank line before class
    "D213",  # multi-line summary second line
    "TD001", # invalid todo tag
    "TD002", # missing todo author
    "TD003", # missing todo link
    "PD011", # pandas use of dot values
    "T201",  # print() の使用を許可
    "D400",  # docstringの末尾がピリオド.の強制を無効化
    "D415",  # docstringの末尾がピリオド.の強制を無効化
    "INP001", # __init__.py の設定不要
    "ERA001", # コメントアウトされたコードをOK
    "RET504", # 直接returnしなくてもOK
    "S311", # random.random使ってもOK
]
unfixable = [
    "F401", # unused import
    "F841", # unused variable
]

logger-objects = ["src.library.logger.LOGGER"]

[tool.ruff.lint.pylint]
max-args = 6

[dependency-groups]
dev = [
    "sqlacodegen>=3.0.0",
]
